+++ CHANGELOG.md
@@ -2,18 +2,23 @@
 - Add StackFit as a parameter
 
 ### [2.0.0+2] - 5 December, 2021
+
 - Use path_drawing: ^1.0.0
 
 ### [2.0.0+1] - 10 August, 2021
+
 - Fixed wrong placement for CircleBorder
 
 ### [2.0.0] - 17 April, 2021
+
 - Move null safety to stable and master
 
 ### [2.0.0-nullsafety.0] - 23 December, 2020
+
 - Migrated package to use null safety
 
 ### [1.0.6] - 16 July, 2020
+
 - Formatted code and added code comments
 
 ### [1.0.5] - 3 March, 2020